+++ pubdoc/lib/src/environment.dart
@@ -21,5 +21,5 @@ class PlatformEnvironment implements Environment {
   String? getVariable(String name) => Platform.environment[name];
 
   @override
-  String get toolVersion => '0.1.2';
+  String get toolVersion => '0.1.3';
 }